+++ CHANGELOG.md
@@ -11,6 +11,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Criado componente `DeleteAccountCard` [#236] (https://github.com/3C-gg/reload-frontend/issues/236).
+- Adicionado método `delete` na interface de AccountsAPI.
+- Adicionado método `updateIsActive` na interface de AccountsAPI.
+- Criado componente `InactivateAccountCard` [#235] (https://github.com/3C-gg/reload-frontend/issues/235).
+- Criado componente `ModalConfirmation`.
+- Adicionado método de `updateEmail` na interface de AccountsAPI.
+- Criado componente `PencilIcon`.
+- Adicionado método de `update` na interface de AccountsAPI.
+- Criado componente `ChangeEmailCard` [#231] (https://github.com/3C-gg/reload-frontend/issues/231).
+- Criado componente `AccountCard`.
+- Adicionado variant `disabled` para input no tema do chakra ui.
 - Adicionado interface `MatchAPI` para controlar as requisições de partidas [#206](https://github.com/3C-gg/reload-frontend/issues/206).
 - Criado interface `ProfilesAPI`.
 - Adicionado tamanho `xl` de `80px` para o componente de `Avatar` no tema do chakra.
@@ -43,6 +54,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Alterado componente `ChangeEmailCard` para usar o método `updateEmail`.
+- Alterado background do `Modal` para `gray.800` no tema do chakra.
+- Alterado gray 800 para `#1B1B1B`.
 - Alterado `HeaderProfile` e `LevelStatsCard` para não quebrar quando o usuário logado for uma conta recém criada.
 - Alterado `HttpService` para adicionar uma `/` no final da url apenas quando não tiver uma interrogação `?` [#243] (https://github.com/3C-gg/reload-frontend/issues/243).
 - Alterado componente `MatchHistoryList` para exibir paginação apenas quando tiver mais que 1 página [#228] (https://github.com/3C-gg/reload-frontend/issues/228).